# docs/swagger.yaml
basePath: /
definitions:
  badresponse.ResponseError:
    properties:
      reason:
        type: string
      time:
        type: string
    type: object
  dto.ImgUploadToken:
    properties:
      file_key:
        type: string
      token:
        type: string
    type: object
  dto.RelateEntity2Entity:
    properties:
      relateTo_id:
        type: string
      related_id:
        type: string
    type: object
  dto.RelateEntitySet2Entity:
    properties:
      relateTo_id:
        type: string
      related_ids:
        items:
          type: string
        type: array
    type: object
...
  model.Comment:
    properties:
      article:
        type: string
      aurhor:
        type: string
      content:
        example: a comment
        type: string
      entity_info:
        $ref: '#/definitions/model.Entity'
      id:
        example: xxxxxxxxxxxxxx==
        type: string
      praise:
        $ref: '#/definitions/model.Praiseable'
      subs:
        items:
          $ref: '#/definitions/model.Subcomment'
        type: array
    type: object
...
  model.Directory:
    properties:
      id:
        example: xxxxxxxxxxxxxx==
        type: string
      info:
        $ref: '#/definitions/model.DirectoryInfo'
      name:
        example: records
        type: string
      owner_ids:
        items:
          type: string
        type: array
      parent_id:
        example: xxxxxxxxxxxxxx==
        type: string
    type: object
  model.DirectoryInfo:
    properties:
      cover:
        example: https://www.motwo.cn/cover
        type: string
      description:
        example: course materials
        type: string
    type: object
  model.Entity:
    properties:
      createTime:
        example: "2020-10-1"
        type: string
      is_deleted:
        type: boolean
      updateTime:
        example: "2020-10-1"
        type: string
    type: object
...
  model.Praiseable:
    properties:
      down:
        type: integer
      up:
        type: integer
      weighted:
        type: integer
    type: object
  model.Subcomment:
    properties:
      aurhor:
        type: string
      content:
        example: a comment
        type: string
      entity_info:
        $ref: '#/definitions/model.Entity'
      id:
        example: xxxxxxxxxxxxxx==
        type: string
      praise:
        $ref: '#/definitions/model.Praiseable'
    type: object
host: localhost:5001
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a Mo2 server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Mo2
  version: "1.0"
paths:
...
  /api/blogs/category:
    get:
      description: 若id为空，返回所有categories(debug mode)；若id不为空，返回该id的category
      parameters:
      - description: string ObjectID
        in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Directory'
            type: array
      summary: find categories
      tags:
      - category
    post:
      consumes:
      - application/json
      description: add by json
      parameters:
      - description: Add category
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/model.Directory'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Directory'
      summary: upsert category
      tags:
      - category
...
  /api/comment:
    post:
      consumes:
      - application/json
      description: upsert json comments
      parameters:
      - description: comment
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/model.Comment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Comment'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/badresponse.ResponseError'
      summary: upsert comments
      tags:
      - comments
  /api/comment/{id}:
    get:
      consumes:
      - application/json
      description: get json comments
      parameters:
      - description: article id
        in: path
        name: id
        required: true
        type: string
      - description: int 0
        in: query
        name: page
        type: integer
      - description: int 5
        in: query
        name: pagesize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Comment'
            type: array
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/badresponse.ResponseError'
      summary: get comments
      tags:
      - comments
    post:
      consumes:
      - application/json
      description: upsert json comments
      parameters:
      - description: comment id
        in: path
        name: id
        required: true
        type: string
      - description: subcomment
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/model.Subcomment'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Subcomment'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/badresponse.ResponseError'
      summary: upsert subcomments
      tags:
      - comments
  /api/commentcount/{id}:
    get:
      description: get article comment num
      parameters:
      - description: article id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: integer
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/badresponse.ResponseError'
      summary: count comments
      tags:
      - comments
...
  /api/directories/category:
    delete:
      consumes:
      - application/json
      description: 根据id删除，并解除与之相关实体的所有关联
      parameters:
      - description: category id to delete
        in: body
        name: ids
        required: true
        schema:
          items:
            type: string
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Directory'
      summary: delete category
      tags:
      - category
...
  /api/img/{filename}:
    get:
      description: add by json
      parameters:
      - description: file name
        in: path
        name: filename
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ImgUploadToken'
      summary: Gen img token
      tags:
      - img
...
  /api/relation/categories/{type}:
    post:
      consumes:
      - application/json
      description: （根据path中提供的关联类型选择对应方法）目前有：父category
      parameters:
      - description: types to relate
        in: path
        name: type
        required: true
        type: string
      - description: sub category id and parent id
        in: body
        name: id
        required: true
        schema:
          $ref: '#/definitions/dto.RelateEntitySet2Entity'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Directory'
      summary: 将列表内的子categories关联到单个实体上
      tags:
      - relation
  /api/relation/category/{type}:
    post:
      consumes:
      - application/json
      parameters:
      - description: relatedTo user/blog/category/userMain
        in: path
        name: type
        required: true
        type: string
      - description: category id to be related
        in: body
        name: e2e
        required: true
        schema:
          $ref: '#/definitions/dto.RelateEntity2Entity'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RelateEntity2Entity'
      summary: relate category to given type
      tags:
      - relation
  /api/relation/category/{type}/{ID}:
    get:
      consumes:
      - application/json
      description: 根据type返回不同结果：[user] 个人的所有category|[sub] 所有子category
      parameters:
      - description: find by user/sub
        in: path
        name: type
        required: true
        type: string
      - description: ID
        in: path
        name: ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Directory'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/badresponse.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/badresponse.ResponseError'
      summary: find categories by given type
      tags:
      - relation
swagger: "2.0"